Consolidate dynamics UGens
